Included are the main validation methods for the following situation:
In VIITE the last road part of a road address should always have Jatkuvuus (continuity) code 1 (end of the road). If road addresses last road part is fully terminated in a road address project (due to operations Lakkautus, Siirto), the previous road part (which is now the last) of the road address should have Jatkuvuus code 1 (end of the road). The previous road part is not necessarily reserved to the road address project. This is why validation check should include road parts outside of the current project.