VIITE-1316
Added a new project state (value = 8) representing a error in Viite.
Added said state to the list of allowed states for a project to be "Writable".
If there is a nonFatal error on Viite, while sending a project to TR then said project state will be updated to this newly created state.
Added the new state state to the list of states that the project window will show.
Also added the new state to the list of publishable project states, in order to allow re-submission on TR part.