VIITE-1361 Made a simple method to return the name of the road that is being split.
The origin is firstly the project link, then the road address, if none if found from these 2 sources then we return the original, un-modified value of the name that resides in the project link.

Also corrected a little bug, resulting of calling a .map() method on a empty collection.

Added the road name parameter to the