VIITE-1262 - Projektilistalle painike "Näytä vain omat projektit" / "Näytä vain omat projektit" (Show only own projects) button to the project list
Modified slightly the backend for getRoles in order to also return the username.
Said username will be extremely useful in order to filter by the current session user immediately.
Modified ProjectListModel in order to add a filter to the project list by the current session user.
Filter is acessed by clicking "Käyttäjä" once the project lists are loaded.