VIITE-1529 - UI bugs after major refactoring
Possible correction of the following bugs:
Blowing up error when the previous selected returned a empty optional.
The unknowns not being drawn on top of floating in theory a defer allows the pause necessary for OL to catch up to the commands sent.
The problem with the clicking on the empty map after being in the defloating process would allow a deselection and enabling of the menus.