VIITE-1526 - Selection corrections
Added names to the layers (useful in testing).
Centralized the affectation to the roadLinkGroups variable (useful in debugging).
Added the ability to find tmpRoads (unknowns associated with Floatings).
Modified the open function in order to findUnknowns in the tmpRoads.
Added a aditional check on the processOL3Features to also check for same anomaly code and same roadLinkType.