VIITE-1601_V2_problems_with_new_two_track_road
Moved distancesBetweenEndPointsInOrigin to GeometryUtils since it is now needed for the determination of the starting point of the recalculate project.
Added a simple getEndPoints that returns what endpoint should be used for (start, end) based on the direction vector supplied.
Changed how to determine the starting points, now it is based on the distance between endpoints in origin and if the right head of geometry is adjacent to the already calculated start point for the right track code.
Changed how the directional vector is determined, if we have no project links with a sideCode != Unknown then the vector will be fixed at (1,1,0) (1st quadrant) or it will be calculated using the old method using only the links with a side code != Unknown.
Added an additional filter at frontEnd that will only emit directional markers if, in addition to the previous rules the project link in question has a side code !== 99(Unknown).