VIITE-1657 - Recalculation in two-tracks does not adjust addr M Values to its original geometry length
In a project with roads with 2 different track codes, the user assigns one value to ALL the tracks (resulting in only having either Left Side XOR Right Side) it will remove the side codes of all the roads involved.
Changed the insert property in to add the reversed property in the new link if ALL THE OTHERS are reversed.
Included the reversed property in the finding of the start points for the calculation.
Added the checking of the reversed property when assigning MAddressValues and SideCodes.