VIITE-1657 - Recalculation in two-tracks does not adjust addr M Values to its original geometry length
Added the passage of the reversed parameter to the backend and it's getting.
Added the ability to check if the calibration point exists at a specific M address value.
Added the correct setting of the reversed parameter on project link creation.
Added an aditional condition that if the project link already has a calibration point then no new one is created on update.