VIITE-1707 The road name does not fill automatically when creating a new road address / Tiennimi ei päivity automaattisesti uutta tieosoitetta luotaessa
Changed the backend to receive 2 parameters, linkId and current Project Id.
Edited the PreFillInfo object to add a parameter that indicates the source of the road name (relevant to the blockage of the name box in the UI).
parsePreFillData now fetches data from either the road name table (this source will block the name box in the UI) or from the project link (this will not block the box).
Added the current project id in the forms to be passed to the backend.
Blockage of the road name box is now dependant of the source.