Viite 3085 date validation for road address changes browser window (#1565)
Road address changes browser window now validates the date inputs for correct format and start date is made compulsory. 

The isFinnishDateString now uses named capture groups in the regex as ESLint warnings suggests.

The switch (targetValue.value) in RoadAddressBrowserWindow.js was still incorrectly using date object as a parameter instead of string value, which is now used with validateDate..

The validateDate now uses isFinnishDateString to check for the correct format of both startDate and endDate. If startDate is not provided the search is denied with "Alkupäivämäärä on pakollinen tieto" validation warning. Added eventListeners for date inputs to empty the validation warnings when typing continues.