VIITE-3120 add project link address (addrMValue) sliding to the recalculation phase.
The addr M values can now "slide" on the project links that are between calibration points and on the same roadway number. This will help the splitting of two track roads and matching the track lengths.

Terminated links will be adjusted if the project links before the terminated links have changed addr M values. The terminated links are returned with all the other links.