VIITE-3096 added a discontinuity validation to one of the address validation cases
If the discontinuity codes are not set properly by the user the code will skip the part where terminated sections are adjusted, which can lead to address mismatch error.

When there are length mismatches the code will first check for discontinuity validation errors and if there are invalid discontinuities the user will be guided towards fixing those issues. If the discontinuity codes are all correct then the user is presented with the length mismatch error (usually the user cannot advance in this situation and technical review of the project and the codebase is needed).